# openapi.yaml
openapi: 3.0.3
info:
  title: Purrfect API
  version: 1.0.0
  description: >
    Secure cat marketplace backend with escrow milestone settlement,
    veterinary inspection window, and dispute moderation.
servers:
  - url: https://api.purrfect.kz
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /listings:
    get:
      tags: [listings]
      summary: List published listings (cursor pagination)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [PUBLISHED, RESERVED, SOLD]
          example: PUBLISHED
        - in: query
          name: breed
          schema:
            type: string
          example: Maine Coon
        - in: query
          name: city
          schema:
            type: string
          example: Almaty
        - in: query
          name: sort
          schema:
            type: string
            enum: [-createdAt, -priceKzt, riskScore]
          example: -createdAt
      responses:
        '200':
          description: Listing page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListings'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [listings]
      summary: Create listing draft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /moderation/listings:
    get:
      tags: [moderation]
      summary: List listing moderation queue
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200': { description: Listing moderation queue }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/listings/{listingId}/approve:
    post:
      tags: [moderation]
      summary: Approve pending listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Listing approved }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/listings/{listingId}/reject:
    post:
      tags: [moderation]
      summary: Reject pending listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Listing rejected }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/listings/{listingId}/risk-flag:
    post:
      tags: [moderation]
      summary: Set listing risk score and create risk case
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Risk flag applied }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/listings/{listingId}/risk-signals:
    get:
      tags: [moderation]
      summary: Get risk signals for listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Listing risk signals }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /moderation/listings/{listingId}/risk-flag/clear:
    post:
      tags: [moderation]
      summary: Clear all OPEN risk flags on a listing and reset risk score
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200':
          description: Risk flags cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  listingId: { type: string, format: uuid }
                  closedCases: { type: integer, example: 2 }
                  currentRiskScore: { type: integer, example: 0 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/listings/{listingId}/full-context:
    get:
      tags: [moderation]
      summary: Full moderation context — listing + media + docs + cases + seller
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200':
          description: Aggregate snapshot for moderator review
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      in: query
      name: cursor
      required: false
      schema:
        type: string
      description: Opaque keyset cursor from previous page
      example: eyJjcmVhdGVkQXQiOiIyMDI2LTA0LTA3VDEzOjQyOjAwWiIsImlkIjoiY2YxMiJ9
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ListingIdParam:
      in: path
      name: listingId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: Invalid query parameter value
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Access token is invalid or expired
    Forbidden:
      description: Authenticated but missing required role/permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Admin role is required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Listing not found
    Conflict:
      description: Business state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Order is already in disputed state
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: inspection report URL is required
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Unexpected server error
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
...
    CreateListingRequest:
      type: object
      required: [title, description, breed, gender, birthDate, vaccinationStatus, priceKzt, city]
      properties:
        title:
          type: string
          example: Scottish Fold kitten, vaccinated
        description:
          type: string
          example: Playful kitten with health passport and first vaccine.
        breed:
          type: string
          example: Scottish Fold
        gender:
          type: string
          enum: [MALE, FEMALE]
          example: FEMALE
        birthDate:
          type: string
          format: date
          example: '2026-01-12'
        vaccinationStatus:
          type: string
          example: 1st dose completed
        priceKzt:
          type: number
          format: double
          example: 250000
        city:
          type: string
          example: Astana
...
    Listing:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        title:
          type: string
        breed:
          type: string
        priceKzt:
          type: number
          format: double
        city:
          type: string
        status:
          type: string
          enum: [DRAFT, PENDING_REVIEW, PUBLISHED, RESERVED, SOLD, REJECTED, ARCHIVED]
        riskScore:
          type: integer
          example: 34
        createdAt:
          type: string
          format: date-time
...
    PageMeta:
      type: object
      required: [hasNext, nextCursor]
      properties:
        hasNext:
          type: boolean
        nextCursor:
          type: string
          nullable: true
    PaginatedListings:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        meta:
          $ref: '#/components/schemas/PageMeta'
...